feat(noscript): Add hook to detect client side rendering, move PIXI stuff out to distinct folder
