+++ src/components/landing/Blobs/PIXIBlobs/circle-64.ico
