+++ src/components/landing/Blobs/styles.js
@@ -8,7 +8,7 @@ export const Wrapper = styled.div`
   background-repeat: no-repeat;
 `;
 
-export const SkillsWrapper = styled.div`
+export const BlobsWrapper = styled.div`
   padding: 4rem 0;
   display: flex;
   align-items: center;
@@ -16,38 +16,5 @@ export const SkillsWrapper = styled.div`
 
   @media (max-width: 960px) {
     flex-direction: column;
-  }
-`;
-
-export const Details = styled.div`
-  flex: 1;
-  padding-left: 2rem;
-  text-align: center;
-
-  @media (max-width: 960px) {
-    padding-left: unset;
-    width: 100%;
-  }
-
-  h1 {
-    margin-bottom: 2rem;
-    font-size: 26pt;
-    color: ${({ theme }) => (theme === 'dark' ? '#fff' : '#212121')};
-
-    @media (max-width: 960px) {
-      mix-blend-mode: ${({ theme }) => (theme === 'light' ? 'unset' : 'difference')};
-    }
-  }
-
-  p {
-    margin-bottom: 2.5rem;
-    font-size: 20pt;
-    font-weight: normal;
-    line-height: 1.3;
-    color: ${({ theme }) => (theme === 'dark' ? '#c7c7c7' : '#707070')};
-
-    @media (max-width: 960px) {
-      mix-blend-mode: ${({ theme }) => (theme === 'light' ? 'unset' : 'difference')};
-    }
   }
 `;
